LayoutTests/fast/dom/Window/open-window-min-size.html

<!-- webkit-test-runner [ ShowModalDialogEnabled=true ] -->
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script>
window.jsTestIsAsync = true;
function validate(msg) {
    shouldBeGreaterThanOrEqual(msg["width"].toString(), "100");
    shouldBeGreaterThanOrEqual(msg["height"].toString(), "100");
}

function test() {
    var data = window.showModalDialog("resources/report-size-and-close.html", "", "dialogWidth:10; dialogHeight:10");
    validate(data);

    window.addEventListener('message', function(e) {
        validate(e.data);
        e.source.close();
        finishJSTest();
    }, false);
    window.open("resources/report-size-and-close.html", "non-empty-argument", "width=10, height=10");
}


</script>
</head>
<body onload="test()">
</body>
</html>
